pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "gbmtsplits"
dynamic = ["version"]
description = "A tool to create well-balanced data splits for multi-task learning"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [{name = "Sohvi Luukkonen", email = "s.luukkonen@lacdr.leidenuniv.nl"},]
license = { file = "LICENSE" }
keywords = ["data split", "multi-task", "balanced data"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.8",
]

dependencies = [
    "parameterized",
    "matplotlib",
    "seaborn",
    "pandas",
    "scikit-learn",
    "rdkit",
    "numpy",
    "tqdm",
    "pulp",
    "biopython"
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools_scm]
write_to = "gbmtsplits/_version.py"
